MSVC errors

+++ src/window.cpp
@@ -1,3 +1,4 @@
+#include <cassert>
 #include "../include/window.h"
 #include "../include/diagnostic.h"
 
@@ -28,8 +29,8 @@ uint16_t
 window_throttle(uint16_t request, uint16_t limit) {
     assert(window_validate(request));
     assert(window_validate(limit));
-    uint8_t request2 = window_apply_default(request);
-    uint8_t limit2 = window_apply_default(limit);
+    uint16_t request2 = window_apply_default(request);
+    uint16_t limit2 = window_apply_default(limit);
     if (request2 < limit2)
         return request2;
     return limit2;
@@ -42,8 +43,8 @@ window_negotiate(uint16_t request, uint16_t current) {
     assert(window_validate(request));
     assert(window_validate(current));
 
-    uint8_t request2 = window_apply_default(request);
-    uint8_t current2 = window_apply_default(current);
+    uint16_t request2 = window_apply_default(request);
+    uint16_t current2 = window_apply_default(current);
     if (current2 >= WINDOW_NOMINAL && (request2 > current2 || request2 < WINDOW_NOMINAL)) {
         diagnostic = d_window_invalid_negotiation;
         return false;
